requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-ckeditor==6.7.3
django-cuser==2017.3.16
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-tinymce==4.1.0
idna==3.10
modeltranslation==0.25
pillow==11.3.0
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
